// src/model/TSet.ts
export interface TSet {
    satzlaenge: string;
    satzart: string;
    blz: string;
    iso_waehrungs: string;
    kontonummers: string;
    ausfuehrungstermin: string;
    blzkonotfuehrend: string;
    iso_waehrunge: string;
    kontonummere: string;
    bic: string;
    lande: string;
    anschrift: string;
    landz: string;
    zahlungsempfaenger: string;
    ordervermerk: string;
    ibane: string;
    auftragswaehrung: string;
    betragvorkomma: string;
    betragnachkomma: string;
    verwendungszweck: string;
    weisungsschluessel1: string;
    weisungsschluessel2: string;
    weisungsschluessel3: string;
    weisungsschluessel4: string;
    zusatzinfoweisungsschluessel: string;
    entgeldregelung: string;
    kennzeichnunguzahlungsart: string;
    variablertext: string;
    nametelefon: string;
    meldeschluessel: string;
    reserve: string;
    erweiterungskennzeichnung: string;

}

export default TSet;